Crux vdev Crux.Cache.Guild

Default Crux.Structs.Channel cache.

Unlike other caches this one splits up to different guild processes handling their data individually.

Functions

child_spec(arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

delete(guild_id)

delete(guild_id :: Crux.Rest.snowflake()) :: :ok | :error

Deletes a guild.

> This will remove all associated channels and emojis from the appropriate caches.

delete(guild_id, data)

delete(guild_id :: Crux.Rest.snowflake(), data :: term()) :: :ok | :error

Deletes a:

Crux.Structs.User (effectively their Crux.Structs.Member and, if applicable, Crux.Structs.VoiceState)

Crux.Structs.Role from the guild

Crux.Structs.Channel from the guild

fetch(guild_id)

fetch(guild_id :: Crux.Rest.snowflake()) ::
  {:ok, Crux.Structs.Guild.t()} | :error

Fetches a guild from the cache by id.

fetch!(guild_id)

fetch!(guild_id :: Crux.Rest.snowflake()) ::
  Crux.Structs.Guild.t() | no_return()

Fetches a guild from the cache by id, raises if not found.

init(guild)

Invoked when the server is started. start_link/3 or start/3 will
block until it returns.

args is the argument term (second argument) passed to start_link/3.

Returning {:ok, state} will cause start_link/3 to return
{:ok, pid} and the process to enter its loop.

Returning {:ok, state, timeout} is similar to {:ok, state}
except handle_info(:timeout, state) will be called after timeout
milliseconds if no messages are received within the timeout.

Returning {:ok, state, :hibernate} is similar to
{:ok, state} except the process is hibernated before entering the loop. See
c:handle_call/3 for more information on hibernation.

Returning :ignore will cause start_link/3 to return :ignore and the
process will exit normally without entering the loop or calling c:terminate/2.
If used when part of a supervision tree the parent supervisor will not fail
to start nor immediately try to restart the GenServer. The remainder of the
supervision tree will be (re)started and so the GenServer should not be
required by other processes. It can be started later with
Supervisor.restart_child/2 as the child specification is saved in the parent
supervisor. The main use cases for this are:

The GenServer is disabled by configuration but might be enabled later.

An error occurred and it will be handled by a different mechanism than the
Supervisor. Likely this approach involves calling Supervisor.restart_child/2
after a delay to attempt a restart.

Returning {:stop, reason} will cause start_link/3 to return
{:error, reason} and the process to exit with reason reason without
entering the loop or calling c:terminate/2.

Callback implementation for GenServer.init/1.

insert(data)

insert(data :: term()) :: term()

Inserts a:

Crux.Structs.Guild itself

Crux.Structs.Member in it

Chunk of Crux.Structs.Member

Crux.Structs.Role

Crux.Structs.Member’s roles

Crux.Structs.VoiceState

lookup(guild_id)

lookup(guild_id :: Crux.Rest.snowflake()) :: {:ok, pid()} | :error

Looks up the pid/0 of a Crux.Cache.Guild’s GenServer by guild id.

update(data)

update(data :: term()) :: term()

Updates or inserts a:

Crux.Structs.Guild itself

Crux.Structs.Member in it

Chunk of Crux.Structs.Member

Crux.Structs.Role

Crux.Structs.Member’s roles

Crux.Structs.VoiceState