Crux vdev Crux.Extensions.Command behaviour

Behaviour module used to compose command pipelines.

Examples

A Simple Ping Command

  defmodule MyBot.Command.Ping do
    use Crux.Extensions.Command

    def triggers(), do: ["ping"]

    def call(command, _opts) do
      set_response(command, content: "Pong!")
    end
  end

A Simple Middleware Command

  defmodule MyBot.Middleware.FetchPicture do
    use Crux.Extensions.Command

    def call(command, opts) do
      # Default the type to :cat
      type = Keyword.get(opts, :type, :cat)

      case MyBot.Api.fetch_picture(type) do
        {:ok, picture} ->
          assign(command, :picture, picture)
        {:error, _error} ->
          command
          |> set_response(content: "An error occurred while fetching the picture.")
          # Halt the pipeline to stop any further execution of commands or middlewares
          |> halt()
      end
    end
  end

Using the Middleware Command

  defmodule MyBot.Command.Dog do
    use Crux.Extensions.Command

    def triggers(), do: ["dog"]

    # Specify the type of :dog
    def required(), do: [{MyBot.Middleware.FetchPicture, type: :dog}]

    def call(command, _opts) do
      set_response(command, content: "Your dog picture link: #{command.assigns.picture}")
    end
  end

Types

command()

command() :: command_mod() | {command_mod(), command_opts()}

A command module, or command module and options tuple.

command_mod()

command_mod() :: module()

A module implementing the Crux.Extensions.Command behaviour.

command_opts()

command_opts() :: term()

Options being passed to the call/2 of a command_mod/0.

The exact type is defined by the command itself.

t()

t() :: %Crux.Extensions.Command{
  args: [String.t()],
  assigns: %{optional(atom()) => term()},
  halted: boolean(),
  message: Crux.Structs.Message.t(),
  response: term(),
  response_channel: Crux.Rest.Util.channel_id_resolvable(),
  shard_id: non_neg_integer(),
  trigger: String.t()
}

Represents the current state of an executing command.

Functions

assign(command, key, value)

assign(t(), key :: atom(), value :: term()) :: t()

Assigns an arbitrary value to an atom key, which is accessible
under the assigns field of a Command.

halt(command)

halt(t()) :: t()

Halts this command, no other commands will be executed fater this one.

set_response(command, response)

set_response(t(), term()) :: t()

Sets the response content for this command.

set_response_channel(command, response_channel)

set_response_channel(t(), Crux.Rest.Util.channel_id_resolvable() | nil) ::
  t()

Sets the response channel for this command.

Callbacks

call(t, command_opts)

call(t(), command_opts()) :: t()

Exeucting this command module.

required() (optional)

required() :: [command_mod() | {command_mod(), command_opts()}]

Returns a list of required command modules to run before this one.

triggers() (optional)

triggers() :: [String.t() | nil]

List of possible triggers for this command module.

Only used and required for primarily handled commands.