Crux v2019-11-21 13:31:04Z Crux.Extensions.Command.Handler behaviour

Handler module serving as entry point for command pipelines.

Example

  defmodule MyBot.Handler do
    use Crux.Extensions.Command.Handler

    def commands() do
      [
        MyBot.Command.Ping,
        MyBot.Command.Dog
      ]
    end
    def producers(), do: MyBot.Producers.fetch()
    def rest(), do: MyBot.Rest
    def prefixes(), do: ["!"]
  end

Callbacks

commands()

commands() :: [Crux.Extensions.Command.command()]

Gets all available commands.

handle_prefixes(message) (optional)

handle_prefixes(message :: Crux.Structs.Message.t()) :: [
  {:ok, String.t()} | :error
]

Removes the prefix from content.

Returns a list of:

{:ok, content} if a prefix matched

:error if a prefix did not

match_commands(content, message, shard_id) (optional)

match_commands(
  content :: String.t(),
  message :: Crux.Structs.Message.t(),
  shard_id :: non_neg_integer()
) :: [{Crux.Extensions.Command.command_info(), Crux.Extensions.Command.t()}]

Gets all commands matching the given content.

on_error(arg1, todo)

on_error(Crux.Extensions.Command.t(), todo :: any()) :: any()

Called when an error occured.

> Not implementing this function causes any errors to terminate the handling process.

prefixes()

prefixes() :: [String.t() | nil]

Gets the current prefixes.

> nil is used for "no prefix"

producers()

producers() :: [pid()]

Gets the current producers.

respond(arg1) (optional)

respond(Crux.Extensions.Command.t()) :: any()

Called after the command is done to send the response.

rest()

rest() :: module()

Gets the module handling rest.

> See Crux.Rest