Crux vdev Crux.Gateway.Connection

Handles the actual connection to Discord.

Functions

send_command(command, shard_id)

send_command(command :: WebSockex.Frame.t(), shard_id :: non_neg_integer()) ::
  :ok | {:error, term()}

Sends a command to the specified shard.

The command will be run through a rate limiter, this blockes the calling process until the command is sent.