Crux vdev Crux.Gateway.Connection.Producer

Handles dispatching of packets received from the gateway.

Every gateway has its own producer, defaults to GenStage.BroadcastDispatchers.

The dispatcher can be overriden via app config or passed override in Crux.Gateway.start/1.
The key is :dispatcher, value should be a valid GenStage.Dispatcher, or a tuple of one and initial state.

> For more informations regarding Consumers and Producers consult GenStage’s docs.

Functions

producers()

producers() :: %{optional(non_neg_integer()) => pid() | :not_found}

Computes a map of all producers keyed by shard_id.

Values are either a pid/0 or, if for some reason the producer could not be found, :not_found.