Crux vdev Crux.Gateway.IdentifyLimiter

Handles the Identifying rate limit of 1 per 5 seconds.

This module is automatically used by Crux.Gateway.Connection, you do not need to worry about it.

Functions

child_spec(arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

queue(packet, shard_id \\ nil)

queue(packet :: term(), shard_id :: non_neg_integer()) :: term()

Starts the module if necessary and queues the packet.
Blocks the process until the identify may be sent.

The packets will return in the order as they arrive at the rate limiter, those are sent via GenServer.call/2.
Returns the packet as is.

Automatically used by Crux.Gateway.Connection.

start_link(args)

Starts the identify limiter.