Crux vdev Crux.Gateway.Util

Collection of util functions.

Both functions are identical to the ones in Crux.Structs.Util.

Functions

atomify(struct)

atomify(input :: struct() | map() | list()) :: map() | list()

Atomifies all keys in a passed list or map to avoid the mess of mixed string and atom keys the gateway sends.

string_to_atom(atom)

string_to_atom(string :: String.t() | atom()) :: atom()

Converts a string to an atom.

Returns an already converted atom as is instead of raising an error.