Crux vdev Crux.Rest.Functions

Module with functions to create Crux.Rest.Request which can be executed.

There are no bangified functions here as no requests to Discord are actually made.

All functions return a Crux.Rest.Request.

Parameters are as described in Crux.Rest behaviour.

Functions

add_guild_member(guild, user, data)

Adds a user to a guild via a provided oauth2 access token with the guilds.join scope.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.add_guild_member/3.

add_guild_member_role(guild, member, role, reason)

Adds a role to a member.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.add_guild_member_role/4.

add_pinned_message(map)

Adds a message to the pinned messages of a channel.

You may only have up to 50 pinned messages per channel.
For more informations see Discord Docs.

Callback implementation for Crux.Rest.add_pinned_message/1.

add_pinned_message(channel, message)

Adds a message to the pinned messages of a channel.

You may only have up to 50 pinned messages per channel.
For more informations see Discord Docs.

Callback implementation for Crux.Rest.add_pinned_message/2.

begin_guild_prune(guild, data)

Prunes members in a guild.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.begin_guild_prune/2.

create_channel_invite(channel, data)

Creates an invite to a channel, see Crux.Rest.create_channel_invite_data/0 for available options.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.create_channel_invite/2.

create_dm(user)

Creates a new dm channel with a user.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.create_dm/1.

create_guild(data)

Callback implementation for Crux.Rest.create_guild/1.

create_guild_ban(guild, user, reason \\ nil)

Bans a user from a guild; The user does not have to be part of the guild.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.create_guild_ban/3.

create_guild_channel(guild, data)

Creates a channel in a guild, see Crux.Rest.create_guild_channel_data/0 for available options.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.create_guild_channel/2.

create_guild_emoji(guild, data)

Creates an emoji in a guild, see Crux.Rest.create_guild_emoji_data/0 for available options.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.create_guild_emoji/2.

create_guild_integration(guild, data)

Attaches an integration from the current user to a guild.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.create_guild_integration/2.

create_guild_role(guild, data)

Creates a role in a guild.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.create_guild_role/2.

create_message(channel_or_message, data)

Sends a message to a channel.

For more informations see Discord Docs

Callback implementation for Crux.Rest.create_message/2.

create_reaction(map, emoji)

Creates a reaction on a message, or increases its count by one.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.create_reaction/2.

create_reaction(channel, message, emoji)

Creates a reaction on a message, or increases its count by one.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.create_reaction/3.

delete_all_reactions(map, emoji)

Deletes all reactions from a message.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.delete_all_reactions/2.

delete_all_reactions(channel, message, emoji)

Deletes all reactions from a message.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.delete_all_reactions/3.

delete_channel(channel, reason \\ nil)

Deletes a channel.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.delete_channel/2.

delete_channel_permissions(channel, target, reason \\ nil)

Deletes an overwrite from a channel.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.delete_channel_permissions/3.

delete_guild(guild)

Deletes a guild, can only be used if the executing user is the owner.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.delete_guild/1.

delete_guild_emoji(guild, emoji, reason \\ nil)

Deletes an emoji from a guild.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.delete_guild_emoji/3.

delete_guild_integration(guild, integration)

Deletes an integration from a guild.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.delete_guild_integration/2.

delete_guild_role(guild, role, reason \\ nil)

Deletes a role in a guild.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.delete_guild_role/3.

delete_invite(code)

Deletes an invite.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.delete_invite/1.

delete_message(map)

Deletes a message

For more informations see Discord Docs.

Callback implementation for Crux.Rest.delete_message/1.

delete_message(channel, message)

Deletes a message

For more informations see Discord Docs.

Callback implementation for Crux.Rest.delete_message/2.

delete_messages(channel, messages)

Deletes 2-100 messages not older than 14 days.

For more informations see Discord Docs

Callback implementation for Crux.Rest.delete_messages/2.

delete_pinned_message(map)

Deletes a message from the pinned messages. This does not delete the message itself.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.delete_pinned_message/1.

delete_pinned_message(channel, message)

Deletes a message from the pinned messages. This does not delete the message itself.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.delete_pinned_message/2.

delete_reaction(message_or_channel, emoji_or_message_id, emoji_or_maybe_user \\ "@me", mayber_user \\ "@me")

Deletes a user from a reaction.

The first argument is optional if a Crux.Structs.Message is provided.

For more informations see Discord Docs 2.

Callback implementation for Crux.Rest.delete_reaction/4.

delete_webhook(user, token \\ nil)

Deletes a webhook

For more information see Discord Docs

Callback implementation for Crux.Rest.delete_webhook/2.

edit_channel_permissions(channel, target, data)

Edits or creates an overwrite for a user, or member.

If an id is provided for :target, :type must be specified in Crux.Rest.edit_channel_permissions_data/0.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.edit_channel_permissions/3.

edit_message(map, data)

Edits a message.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.edit_message/2.

edit_message(channel, message, data)

Edits a message.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.edit_message/3.

execute_github_webhook(map, event, data)

execute_github_webhook(user, token, event, wait \\ false, data)

Callback implementation for Crux.Rest.execute_github_webhook/5.

execute_slack_webhook(map, data)

execute_slack_webhook(user, token, wait \\ false, data)

Callback implementation for Crux.Rest.execute_slack_webhook/4.

execute_webhook(map, data)

execute_webhook(user, token, wait \\ false, data)

Callback implementation for Crux.Rest.execute_webhook/4.

gateway()

Gets the gateway url from the api.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.gateway/0.

gateway_bot()

Gets the gateway url along a recommended shards count from the api.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.gateway_bot/0.

get_audit_logs(guild, data \\ [])

Gets the audit logs for a guild

Callback implementation for Crux.Rest.get_audit_logs/2.

get_channel(channel)

Gets a channel from the api.
This should NOT be necessary.

For more informations see Discord Docs

Callback implementation for Crux.Rest.get_channel/1.

get_channel_invites(channel)

Gets invites for the specified channel from the api.

For more informations see Discord Docs

Callback implementation for Crux.Rest.get_channel_invites/1.

get_current_user_guilds(data)

Gets a list of partial Crux.Structs.Guilds the current user is a member of.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.get_current_user_guilds/1.

get_guild(guild)

Gets a guild from the api.
This should usually, due to cache, NOT be necessary.

For more informations see Discord Docs

Callback implementation for Crux.Rest.get_guild/1.

get_guild_ban(guild, user)

Gets a single ban entry by id.

Returns {:error, %Crux.Rest.ApiError{status_code: 404, code: 10026, ...}} when the user is not banned.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.get_guild_ban/2.

get_guild_bans(guild)

Gets a map of banned users along their ban reasons.

For more informations see discord Docs.

Callback implementation for Crux.Rest.get_guild_bans/1.

get_guild_channels(guild)

Gets all channels from a guild via the api.
This should usually, due to caching, NOT be necessary.

For more informations see Discord Docs-

Callback implementation for Crux.Rest.get_guild_channels/1.

get_guild_embed(guild)

Gets a guild's embed (server widget).

Returns a Guild Embed Object.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.get_guild_embed/1.

get_guild_emoji(guild, emoji)

Gets an emoji from a guild
This should usually, due to cache, NOT be necessary.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.get_guild_emoji/2.

get_guild_integrations(guild)

Gets a list of guild integrations.

Returns a list of Integration Objects.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.get_guild_integrations/1.

get_guild_invites(guild)

Gets all available invites in a guild.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.get_guild_invites/1.

get_guild_member(guild, user)

Gets a member from the api.

This may be necessary for offline members in large guilds.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.get_guild_member/2.

get_guild_prune_count(guild, days)

Gets the number of members in a guild that would be removed when pruned.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.get_guild_prune_count/2.

get_guild_roles(guild)

Gets a list of roles in a guild.
This should usually, due to caching, NOT be necessary.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.get_guild_roles/1.

get_guild_vanity_url(guild)

Gets the vanity url of a guild, if any

Callback implementation for Crux.Rest.get_guild_vanity_url/1.

get_guild_voice_regions(guild)

Gets a list of voice regions for a guild. Returns VIP servers when the guild is VIP-enabled.

Returns a list of Voice Region Objects.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.get_guild_voice_regions/1.

get_invite(code)

Gets an invite from the api.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.get_invite/1.

get_message(message_or_channel, data_or_channel \\ [], data \\ [])

get_messages(channel, data)

Gets 1-100 messages from the api, this limit is enforced on discord's end.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.get_messages/2.

get_pinned_messages(channel)

Gets a list of pinned messages from the api.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.get_pinned_messages/1.

get_reactions(channel_or_message, emoji_or_message_id, emoji_or_maybe_data \\ [], maybe_data \\ [])

Gets users who reacted to a message.

The first argument is optional if a Crux.Structs.Message is provided.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.get_reactions/4.

get_user(user)

Gets a user from the api.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.get_user/1.

get_webhook(user, token \\ nil)

Gets a webhook

For more information see Discord Docs

Callback implementation for Crux.Rest.get_webhook/2.

leave_guild(guild)

Leaves a guild.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.leave_guild/1.

list_channel_webhooks(channel)

Gets a channel's webhook list

For more information see Discord Docs

Callback implementation for Crux.Rest.list_channel_webhooks/1.

list_guild_emojis(guild)

Gets a list of emojis in a guild.
This should usually, due to cache, NOT be necessary.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.list_guild_emojis/1.

list_guild_members(guild, options)

Gets a list of members from the guild.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.list_guild_members/2.

list_guild_webhooks(guild)

Gets a guild's webhook list

For more information see Discord Docs

Callback implementation for Crux.Rest.list_guild_webhooks/1.

modify_channel(channel, data)

Modifies a channel, see Crux.Rest.modify_channel_data/0 for available options.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.modify_channel/2.

modify_current_user(data)

Modifes the currently logged in user.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.modify_current_user/1.

modify_current_users_nick(guild, nick, reason)

Modifies the nickname of the current user in a guild.

Yes, you read correctly, that has its own endpoint.
Great, isn't it?

For more informations, but not an answer to the question why, see Discord Docs.

Callback implementation for Crux.Rest.modify_current_users_nick/3.

modify_guild(guild, data)

Updates a guild, see Crux.Rest.modify_guild_data/0 for available options.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.modify_guild/2.

modify_guild_channel_positions(guild, channels)

Modifyies the position of a list of channels in a guild.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.modify_guild_channel_positions/2.

modify_guild_embed(guild, data)

Modifies a guild's embed (server widget).

Returns the updated Guild Embed Object.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.modify_guild_embed/2.

modify_guild_emoji(guild, emoji, data)

Modifies a guild emoji.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.modify_guild_emoji/3.

modify_guild_integration(guild, integration, data)

Modifies an integreation for a guild.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.modify_guild_integration/3.

modify_guild_member(guild, member, data)

Modifies a member in a guild.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.modify_guild_member/3.

modify_guild_role(guild, role, data)

Modifies a role in a guild.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.modify_guild_role/3.

modify_guild_role_positions(guild, data)

Modifies the positions of a list of role objects for a guild.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.modify_guild_role_positions/2.

remove_guild_ban(guild, user, reason \\ nil)

Removes a ban for a user from a guild.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.remove_guild_ban/3.

remove_guild_member_role(guild, member, role, reason \\ nil)

Removes a role from a member.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.remove_guild_member_role/4.

sync_guild_integration(guild, integration)

Syncs an integration for a guild.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.sync_guild_integration/2.

trigger_typing(channel)

Lets the bot appear as typing for roughly ~9 seconds or until a message is sent.
Should generally be used sparingly for commands that may take a while as a form of acknowledging.

Consider sending a message and edit that later on instead.

For more informations see Discord Docs.

Callback implementation for Crux.Rest.trigger_typing/1.

update_webhook(user, token \\ nil, data)

Updates a webhook

For more information see Discord Docs

Callback implementation for Crux.Rest.update_webhook/3.