Crux vdev Crux.Rest.Handler.Global

Handles global rate limits and the average time offset to discord.

All functions in Crux.Rest automatically use this module, you do not need to worry about it.

Functions

add_offset(offset)

add_offset(offset :: integer()) :: :ok

Adds an offset to the list of offsets.

child_spec(arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

fetch_global_wait()

fetch_global_wait() :: integer()

Fetches how long to wait until the global rate limit is over.
Is not positive when not rate limited.

fetch_offset()

fetch_offset() :: integer()

Fetches the average out of the last ten offsets to discord servers in milliseconds.

set_global_wait(retry_after)

set_global_wait(retry_after :: pos_integer()) :: :ok

Sets the global wait time if rate limited globally.

start_link(args)

start_link(route :: String.t()) :: GenServer.on_start()

Starts the global rate limite handler.