Crux vdev Crux.Structs

Provides a unified function to create one or a list of structs, invoking their create/1 function if available.

Functions

create(data, target)

create(data :: map() | list(), target :: module()) :: struct() | list()

Creates a struct or a list of structs invoking their create/1 function if available.