Crux vdev Crux.Structs.AuditLogChange

Represents a Discord Audit Log Change Object

Types

audit_log_change_value() (since 0.1.6)

audit_log_change_value() ::
  String.t()
  | Crux.Rest.snowflake()
  | integer()
  | boolean()
  | [Crux.Structs.Role.t()]
  | [Crux.Structs.Overwrite.t()]

Represents a value before or after a change.

Note that the Role object returned by Discord in audit logs is a partial role that only contains id and name.

t() (since 0.1.6)

t() :: %Crux.Structs.AuditLogChange{
  key: String.t(),
  new_value: audit_log_change_value() | nil,
  old_value: audit_log_change_value() | nil
}

Functions

create(data) (since 0.1.6)

create(data :: map()) :: t()

Creates a Crux.Structs.AuditLogChange struct from raw data.

Automatically invoked by Crux.Structs.create/2.