Crux vdev Crux.Structs.Invite

Represents a Discord Invite Object

List of what property can be present fetched with what function:

Property	Rest.get_invite/1	Rest.create_channel_invite/1	Rest.delete_invite/1

Rest.get_channel_invites/1

Rest.get_guild_invites/1

code	yes	yes	yes

guild	yes	yes	yes

channel	yes	yes	yes

inviter	yes	yes	yes

uses	no	yes	no

max_uses	no	yes	no

max_age	no	yes	no

temporary	no	yes	no

created_at	no	yes	no

revoked	no	no	no

approximate_presence_count	yes	no	no

approximate_member_count	yes	no	no

Notes:

:guild only has:verification_level, :features, :name, :splash, :id, and :icon.

:channel only has :type, :id and :name.

You can, if applicable, fetch the full structs from cache.

Types

t()

t() :: %Crux.Structs.Invite{
  approximate_member_count: integer() | nil,
  approximate_presence_count: integer() | nil,
  channel: Crux.Structs.Channel.t(),
  code: String.t(),
  created_at: String.t() | nil,
  guild: Crux.Structs.Guild.t(),
  inviter: Crux.Structs.User.t() | nil,
  max_age: integer() | nil,
  max_uses: integer() | nil,
  revoked: boolean() | nil,
  temporary: boolean() | nil,
  uses: integer() | nil
}

Functions

create(data)

Creates a Crux.Structs.Invite struct from raw data.

Automatically invoked by Crux.Structs.create/2.