Crux v2019-11-19 15:31:22Z Crux.Structs.Overwrite

Represents a Discord Overwrite Object.

Types

t() (since 0.1.0)

t() :: %Crux.Structs.Overwrite{
  allow: integer(),
  deny: integer(),
  id: Crux.Structs.Snowflake.t(),
  type: String.t()
}

target_resolvable() (since 0.2.1)

target_resolvable() ::
  Crux.Structs.Overwrite.t()
  | Crux.Structs.Role.t()
  | Crux.Structs.User.id_resolvable()

All available types that can be resolved into a target for a permission overwrite

Functions

create(data) (since 0.1.0)

create(data :: map()) :: t()

Creates a Crux.Structs.Overwrite.t/0 struct from raw data.

Automatically invoked by Crux.Structs.create/2.

resolve_target(resolvable)

resolve_target(target_resolvable()) ::
  {String.t() | :unknown, Crux.Structs.Snowflake.t()}

Resolves a target_resolvable/0 into an overwrite target.

Note that an id or string of it returns :unknown as type.

Examples

  iex> %Crux.Structs.Overwrite{type: "member", id: 218348062828003328}
  ...> |> Crux.Structs.Overwrite.resolve_target()
  {"member", 218348062828003328}

  iex> %Crux.Structs.Role{id: 376146940762783746}
  ...> |> Crux.Structs.Overwrite.resolve_target()
  {"role", 376146940762783746}

  iex> %Crux.Structs.User{id: 218348062828003328}
  ...> |> Crux.Structs.Overwrite.resolve_target()
  {"member", 218348062828003328}

  iex> %Crux.Structs.Member{user: 218348062828003328}
  ...> |> Crux.Structs.Overwrite.resolve_target()
  {"member", 218348062828003328}

  iex> %Crux.Structs.Message{author: %Crux.Structs.User{id: 218348062828003328}}
  ...> |> Crux.Structs.Overwrite.resolve_target()
  {"member", 218348062828003328}

  iex> %Crux.Structs.VoiceState{user_id: 218348062828003328}
  ...> |> Crux.Structs.Overwrite.resolve_target()
  {"member", 218348062828003328}

  iex> 218348062828003328
  ...> |> Crux.Structs.Overwrite.resolve_target()
  {:unknown, 218348062828003328}

  iex> "218348062828003328"
  ...> |> Crux.Structs.Overwrite.resolve_target()
  {:unknown, 218348062828003328}

  iex> nil
  ...> |> Crux.Structs.Overwrite.resolve_target()
  nil