Crux vdev Crux.Structs.Permissions

Custom non discord api struct to help with working with permissions.

Types

name()

name() ::
  :create_instant_invite
  | :kick_members
  | :ban_members
  | :administrator
  | :manage_channels
  | :manage_guild
  | :add_reactions
  | :view_audit_log
  | :priority_speaker
  | :view_channel
  | :send_messages
  | :send_tts_message
  | :manage_messages
  | :embed_links
  | :attach_files
  | :read_message_histroy
  | :mention_everyone
  | :use_external_emojis
  | :connect
  | :speak
  | :mute_members
  | :deafen_members
  | :move_members
  | :use_vad
  | :change_nickname
  | :manage_nicknames
  | :manage_roles
  | :manage_webhooks
  | :manage_emojis

Union type of all valid permission name atoms.

resolvable()

resolvable() :: t() | non_neg_integer() | name() | [resolvable()]

All valid types which can be directly resolved into a permissions bitfield.

t()

t() :: %Crux.Structs.Permissions{bitfield: non_neg_integer()}

Represents a Crux.Structs.Permissions.

:bitfield: The raw bitfield of permission flags.

Functions

add(base, to_add)

add(base :: resolvable(), to_add :: resolvable()) :: t()

Adds permissions to the base permissions.

Examples

iex> :administrator
...> |> Crux.Structs.Permissions.add(:manage_guild)
%Crux.Structs.Permissions{bitfield: 0x28}

all()

all() :: pos_integer()

Returns the integer value of all permissions summed up.

flags()

flags() :: %{optional(name()) => non_neg_integer()}

Returns a map of all permissions.

from(member, guild, channel \\ nil)

from(
  member ::
    Crux.Structs.Member.t() | Crux.Structs.User.t() | Crux.Rest.snowflake(),
  guild :: Crux.Structs.Guild.t(),
  channel :: Crux.Structs.Channel.t() | nil
) :: t()

Resolves permissions for a user in a guild, optionally including channel permission overwrites.

Raises when the member is not cached.

The administrator flag or being owner implicitly grants all permissions.

has(have, want, implicit \\ false)

has(have :: resolvable(), want :: resolvable(), implicit :: boolean()) ::
  boolean()

Check whether the second permissions are all present in the first.

The administrator flag optionally implicitly grants all permissions.

Examples

# If implicit is set, administrator will grant all permissions
iex> Crux.Structs.Permissions.has(0x8, 0x7ff7fcf7, true)
true

# If implicit is not set, administrator won't grant any other permissions
iex> Crux.Structs.Permissions.has(0x8, 0x7ff7fcf7)
false

# Resolving a list of `permissions_name`s
iex> Crux.Structs.Permissions.has([:send_messages, :view_channel, :read_message_history], [:send_messages, :view_channel])
true

# Resolving different types of `permissions`s
iex> Crux.Structs.Permissions.has(:administrator, 0x8)
true

# In different order
iex> Crux.Structs.Permissions.has(0x8, :administrator)
true

names()

names() :: [name()]

Returns a list of all permission keys.

new(permissions)

new(permissions :: resolvable()) :: t()

Creates a new Crux.Structs.Permissions struct from a valid t:permissions/0.

remove(base, to_remove)

remove(base :: resolvable(), to_remove :: resolvable()) :: t()

Removes permissions from the base permissions

Examples

iex> [0x8, 0x10, 0x20]
...> |> Crux.Structs.Permissions.remove([0x10, 0x20])
%Crux.Structs.Permissions{bitfield: 0x8}

resolve(permissions)

resolve(permissions :: resolvable()) :: non_neg_integer()

Resolves a resolvable/0 into a bitfield representing the set permissions.

Examples

# A single bitflag
iex> 0x8
...> |> Crux.Structs.Permissions.resolve()
0x8

# A single name
iex> :administrator
...> |> Crux.Structs.Permissions.resolve()
0x8

# A list of bitflags
iex> [0x8, 0x4]
...> |> Crux.Structs.Permissions.resolve()
0xC

# A list of names
iex> [:administrator, :ban_members]
...> |> Crux.Structs.Permissions.resolve()
0xC

# A mixture of both
iex> [:manage_roles, 0x400, 0x800, :add_reactions]
...> |> Crux.Structs.Permissions.resolve()
0x10000C40

to_list(permissions, implicit \\ false)

to_list(permissions :: resolvable(), implicit :: boolean()) :: [name()]

Serializes permissions into a list of set name/0s.

The administrator flag optionally implicitly grants all permissions.

Examples

iex> 0x30
...> |> Crux.Structs.Permissions.to_list()
[:manage_guild, :manage_channels]

to_map(permissions, implicit \\ false)

to_map(permissions :: resolvable(), implicit :: boolean()) :: %{
  optional(name()) => boolean()
}

Serializes permissions into a map keyed by name/0 with a boolean indicating whether the name is set.

The administrator flag optionally implicitly grants all permissions.