Crux v2019-09-13 17:31:47Z Crux.Cache behaviour

Behaviour all caches must implement. (Looking at custom ones you may want to write)

There are exceptions:

User cache:

Implement a me/1 function setting the own user id

A me/0 and me!/0 function getting the own user

Guild cache:

A bit more, you probably want to take a look at the code of the Crux.Cache.Guild module

Custom caches should be put under a Crux.Cache.Provider. (Can be combined with default caches)

Also worth a look:

Crux.Cache.None - A dummy Crux.Cache and Crux.Cache.Provider, not caching anything.

Types

key()

key() :: non_neg_integer()

Default caches are using Discord Snowflakes as identifiers.

Callbacks

delete(id)

delete(id :: key()) :: :ok

Deletes data from the cache by key.

fetch(id)

fetch(id :: key()) :: {:ok, term()} | :error

Fetches data from the cache by key.

fetch!(id)

fetch!(id :: identifier()) :: term() | no_return()

Fetches data from the cache by key, raises if not found.

insert(data)

insert(data :: term()) :: term()

Inserts data into the cache.

Returns the atomified data allowing the operation to be chained.

For example something like that:

id =
  raw_data
  |> Cache.insert()
  |> Map.get(:id)

start_link(args) (optional)

start_link(args :: term()) :: Supervisor.on_start()

Used to start anything fitting under a supervision tree, like for example a GenServer, instructed with handling the cache.

Optional, you maybe want to use external caching, e.g. Redis, not requiring anything like that.

update(data)

update(data :: term()) :: term()

Inserts data into the cache.

Returns "updated" data including changes by merging.
For example from a message embed update to a full message object

content =
  partial_message # only contains `:id`, `:channel_id`, and `:embeds`
  |> Cache.update()
  |> Map.get(:content) # present if the message was cached previously