Crux v2019-11-09 18:57:33Z Crux.Cache.Guild

Default Crux.Structs.Channel cache.

Unlike other caches this one splits up to different guild processes handling their data individually.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

delete(data_or_id)

delete(
  data_or_id ::
    Crux.Rest.snowflake()
    | Crux.Structs.Guild.t()
    | Crux.Structs.Channel.t()
    | Crux.Structs.Role.t()
    | Crux.Structs.Member.t()
) :: :ok

Deletes a:

Crux.Structs.Member (if applicable, also their Crux.Structs.VoiceState) from the guild

Crux.Structs.Role from the guild

Crux.Structs.Channel from the guild

Crux.Structs.Guild itself

This will remove all associated channels and emojis from the appropriate caches.

fetch(guild_id)

fetch(guild_id :: Crux.Rest.snowflake()) ::
  {:ok, Crux.Structs.Guild.t()} | :error

Fetches a guild from the cache by id.

fetch!(guild_id)

fetch!(guild_id :: Crux.Rest.snowflake()) ::
  Crux.Structs.Guild.t() | no_return()

Fetches a guild from the cache by id, raises if not found.

insert(data)

insert(data :: term()) :: term()

Inserts a:

Crux.Structs.Guild itself

Crux.Structs.Member in it

Chunk of Crux.Structs.Member

Crux.Structs.Role

Crux.Structs.Member's roles

Crux.Structs.VoiceState

lookup(guild_id)

lookup(guild_id :: Crux.Rest.snowflake()) :: {:ok, pid()} | :error

Looks up the pid/0 of a Crux.Cache.Guild's GenServer by guild id.

update(data)

update(data :: term()) :: term()

Updates or inserts a:

Crux.Structs.Guild itself

Crux.Structs.Member in it

Chunk of Crux.Structs.Member

Crux.Structs.Role

Crux.Structs.Member's roles

Crux.Structs.VoiceState