Crux v2019-11-09 18:53:50Z Crux.Cache.Provider behaviour

A behaviour module for cache providers.

Caches with a start_link function will be automatically put under the supervision tree when starting the module implementing this behaviour.

Callbacks

channel_cache()

channel_cache() :: module()

Fetches the module handling the channel cache.

emoji_cache()

emoji_cache() :: module()

Fetches the module handling the channel cache.

guild_cache()

guild_cache() :: module()

Fetches the module handling the guild cache.

presence_cache()

presence_cache() :: module()

Fetches the module handling the presence cache.

user_cache()

user_cache() :: module()

Fetches the module handling the user cache.