Crux vdev Crux.Rest.Endpoints

The endpoints being used by the Crux.Rest module, you do not need to worry about it.

Functions

api()

api() :: String.t()

Base API address.

cdn()

cdn() :: String.t()

Base CDN address.

channel(channel_id, suffix \\ nil)

channel(channel_id :: Crux.Rest.snowflake(), suffix :: String.t() | nil) ::
  String.t()

Used for channel related functions.

channel_messages(channel_id, suffix \\ nil)

channel_messages(
  channel_id :: Crux.Rest.snowflake(),
  suffix :: String.t() | nil
) :: String.t()

Used for channel messages related functions.

channel_permissions(channel_id, target_id)

channel_permissions(
  channel_id :: Crux.Rest.snowflake(),
  target_id :: Crux.Rest.snowflake()
) :: String.t()

Used for channel permissions related functions.

channel_pins(channel_id, suffix \\ nil)

channel_pins(channel_id :: Crux.Rest.snowflake(), suffix :: String.t() | nil) ::
  String.t()

Used for pin related functions.

gateway()

gateway() :: String.t()

Used to obtain the gateway address.

gateway_bot()

gateway_bot() :: String.t()

Used to obtain the gateway address along the recommended shard count.

guild(guild_id \\ nil, suffix \\ nil)

guild(guild_id :: Crux.Rest.snowflake(), suffix :: String.t() | nil) ::
  String.t()

Used for guild related functions.

guild_bans(guild_id, suffix \\ nil)

guild_bans(guild_id :: Crux.Rest.snowflake(), suffix :: String.t() | nil) ::
  String.t()

Used for ban related functions.

guild_emojis(guild_id, suffix \\ nil)

guild_emojis(guild_id :: Crux.Rest.snowflake(), suffix :: String.t() | nil) ::
  String.t()

Used for guild emoji related functions.

guild_integrations(guild_id, suffix \\ nil)

guild_integrations(
  guild_id :: Crux.Rest.snowflake(),
  suffix :: String.t() | nil
) :: String.t()

Used for integration related functions.

guild_member_roles(guild_id, member_id, role_id \\ nil)

guild_member_roles(
  guild_id :: Crux.Rest.snowflake(),
  member_id :: Crux.Rest.snowflake(),
  role_id :: Crux.Rest.snowflake()
) :: String.t()

Used for role related functions.

guild_members(guild_id, suffix \\ nil)

guild_members(guild_id :: Crux.Rest.snowflake(), suffix :: String.t() | nil) ::
  String.t()

Used for guild members related functions.

guild_own_nick(guild_id)

guild_own_nick(guild_id :: Crux.Rest.snowflake()) :: String.t()

Discord being special.

guild_roles(guild_id, suffix \\ nil)

guild_roles(guild_id :: Crux.Rest.snowflake(), suffix :: String.t() | nil) ::
  String.t()

Used for role related functions.

invite(code \\ nil)

invite(code :: String.t() | nil) :: String.t()

Used for invite related functions.

message_reactions(channel_id, message_id, emoji, suffix \\ nil)

message_reactions(
  channel_id :: Crux.Rest.snowflake(),
  message_id :: Crux.Rest.snowflake(),
  emoji :: String.t(),
  suffix :: String.t() | nil
) :: String.t()

Used for reactions related functions.