Crux v2019-09-13 17:31:47Z Crux.Rest.Endpoints

Endpoints being used by the Crux.Rest module, you do not need to worry about it.

Functions

base_url() (since 0.1.5)

base_url() :: String.t()

Base API address.

channel(channel_id, suffix \\ nil) (since 0.1.0)

channel(channel_id :: Crux.Rest.snowflake(), suffix :: String.t() | nil) ::
  String.t()

Used for channel related functions.

channel_invites(channel_id) (since 0.2.0)

channel_invites(channel_id :: Crux.Rest.snowflake()) :: String.t()

Used for channel invites.

channel_messages(channel_id, suffix \\ nil) (since 0.1.0)

channel_messages(
  channel_id :: Crux.Rest.snowflake(),
  suffix :: Crux.Rest.snowflake() | String.t() | nil
) :: String.t()

Used for channel messages related functions.

channel_messages_bulk_delete(channel_id) (since 0.2.0)

channel_messages_bulk_delete(channel_id :: Crux.Rest.snowflake()) :: String.t()

Used for message deletion functions.

channel_permissions(channel_id, target_id) (since 0.1.0)

channel_permissions(
  channel_id :: Crux.Rest.snowflake(),
  target_id :: Crux.Rest.snowflake()
) :: String.t()

Used for channel permissions related functions.

channel_pins(channel_id, suffix \\ nil) (since 0.1.0)

channel_pins(channel_id :: Crux.Rest.snowflake(), suffix :: String.t() | nil) ::
  String.t()

Used for pin related functions.

channel_typing(channel_id) (since 0.2.0)

channel_typing(channel_id :: Crux.Rest.snowflake()) :: String.t()

Used for channel typing.

channel_webhooks(channel_id) (since 0.1.7)

channel_webhooks(channel_id :: Crux.Rest.snowflake()) :: String.t()

Used to fetch channel webhooks.

gateway() (since 0.1.0)

gateway() :: String.t()

Used to obtain the gateway address.

gateway_bot() (since 0.1.0)

gateway_bot() :: String.t()

Used to obtain the gateway address along the recommended shard count.

guild(guild_id \\ nil, suffix \\ nil) (since 0.1.0)

guild(guild_id :: Crux.Rest.snowflake() | nil, suffix :: String.t() | nil) ::
  String.t()

Used for guild related functions.

guild_audit_logs(guild_id) (since 0.1.7)

guild_audit_logs(guild_id :: Crux.Rest.snowflake()) :: String.t()

Used for guild audit logs fetching.

guild_bans(guild_id, suffix \\ nil) (since 0.1.2)

guild_bans(
  guild_id :: Crux.Rest.snowflake(),
  suffix :: Crux.Rest.snowflake() | String.t() | nil
) :: String.t()

Used for ban related functions.

guild_channels(guild_id) (since 0.2.0)

guild_channels(guild_id :: Crux.Rest.snowflake()) :: String.t()

Used for guild channel functions.

guild_embed(guild_id) (since 0.2.0)

guild_embed(guild_id :: Crux.Rest.snowflake()) :: String.t()

Used for guild embed functions.

guild_emojis(guild_id, suffix \\ nil) (since 0.1.0)

guild_emojis(
  guild_id :: Crux.Rest.snowflake(),
  suffix :: Crux.Rest.snowflake() | String.t() | nil
) :: String.t()

Used for guild emoji related functions.

guild_integrations(guild_id, suffix \\ nil) (since 0.1.2)

guild_integrations(
  guild_id :: Crux.Rest.snowflake(),
  suffix :: String.t() | nil
) :: String.t()

Used for integration related functions.

guild_invites(guild_id) (since 0.2.0)

guild_invites(guild_id :: Crux.Rest.snowflake()) :: String.t()

Used for guild invites functions.

guild_member_roles(guild_id, member_id, role_id \\ nil) (since 0.1.1)

guild_member_roles(
  guild_id :: Crux.Rest.snowflake(),
  member_id :: Crux.Rest.snowflake(),
  role_id :: Crux.Rest.snowflake() | nil
) :: String.t()

Used for role related functions.

guild_members(guild_id, suffix \\ nil) (since 0.1.0)

guild_members(
  guild_id :: Crux.Rest.snowflake(),
  suffix :: Crux.Rest.snowflake() | String.t() | nil
) :: String.t()

Used for guild members related functions.

guild_own_nick(guild_id) (since 0.1.0)

guild_own_nick(guild_id :: Crux.Rest.snowflake()) :: String.t()

Discord being special.

guild_prune(guild_id) (since 0.2.0)

guild_prune(guild_id :: Crux.Rest.snowflake()) :: String.t()

Used for guild prune functions

guild_regions(guild_id) (since 0.2.0)

guild_regions(guild_id :: Crux.Rest.snowflake()) :: String.t()

Used for guild region functions.

guild_roles(guild_id, suffix \\ nil) (since 0.1.2)

guild_roles(
  guild_id :: Crux.Rest.snowflake(),
  suffix :: Crux.Rest.snowflake() | String.t() | nil
) :: String.t()

Used for role related functions.

guild_webhooks(guild_id) (since 0.1.7)

guild_webhooks(guild_id :: Crux.Rest.snowflake()) :: String.t()

Used to fetch guild webhooks.

invite(code \\ nil) (since 0.1.0)

invite(code :: String.t() | nil) :: String.t()

Used for invite related functions.

me(suffix \\ nil) (since 0.1.4)

me(suffix :: String.t() | nil) :: String.t()

Used for functions related to the current user.

me_channels() (since 0.2.0)

me_channels() :: String.t()

Used for dm channel functions.

me_guilds(suffix \\ nil) (since 0.1.4)

me_guilds(suffix :: Crux.Rest.snowflake() | String.t() | nil) :: String.t()

Used for functions related to the current user's guilds.

message_reactions(channel_id, message_id, emoji, suffix \\ nil) (since 0.1.0)

message_reactions(
  channel_id :: Crux.Rest.snowflake(),
  message_id :: Crux.Rest.snowflake(),
  emoji :: String.t(),
  suffix :: Crux.Rest.snowflake() | String.t() | nil
) :: String.t()

Used for reactions related functions.

users(suffix \\ nil) (since 0.1.4)

users(suffix :: Crux.Rest.snowflake() | String.t() | nil) :: String.t()

Used for functions related to users.

webhook(webhook_id, token \\ nil) (since 0.1.7)

webhook(guild_id :: Crux.Rest.snowflake(), token :: String.t() | nil) ::
  String.t()

Used for webhook related functions.

webhook_github(webhook_id, token) (since 0.1.7)

webhook_github(webhook_id :: Crux.Rest.snowflake(), token :: String.t()) ::
  String.t()

Used to send github formatted webhooks.

webhook_slack(webhook_id, token) (since 0.1.7)

webhook_slack(webhook_id :: Crux.Rest.snowflake(), token :: String.t()) ::
  String.t()

Used to send slack formatted webhooks.