Crux vdev Crux.Rest.Util

Collection of util functions.

Types

resolved_file()

resolved_file() ::
  {String.t() | :file, String.t() | binary(),
   {String.t(), [{String.t(), binary()}]}, [{String.t(), String.t()}]}
  | {:error, term()}

Used when sending files via Rest.create_message/2.

Functions

map_file(path)

map_file(
  path ::
    String.t()
    | {String.t() | binary(), String.t()}
    | {String.t() | :file, String.t() | binary(), String.t()}
) :: resolved_file() | {:error, term()}

Resolves a:

path to a file

tuple of path to a file or binary of one, and a file name
to a resolved_file automatically used by Rest.create_message/2

resolve_channel_position(arg1)

resolve_channel_position(
  channel ::
    Crux.Structs.Channel.t()
    | map()
    | {integer(), integer() | %{id: integer(), position: integer()}}
) :: %{id: integer(), position: integer()}

resolve_emoji_identifier(emoji)

resolve_emoji_identifier(
  emoji :: Crux.Structs.Reaction.t() | Crux.Structs.Emoji.t() | String.t()
) :: String.t()

resolve_file(file)

resolve_file(file :: String.t() | binary()) ::
  {:ok, binary()} | {:error, term()}

Resolves a string or a binary to a binary/0.

http / https url

local file path

a binary itself

resolve_message_id(id)

resolve_message_id(message :: Crux.Structs.Message.t() | integer()) :: integer()

resolve_overwrite_target(id)

resolve_overwrite_target(
  overwrite ::
    Crux.Structs.Overwrite.t()
    | Crux.Structs.Role.t()
    | Crux.Structs.User.t()
    | Crux.Structs.Member.t()
    | integer()
) :: {:role | :member | :unknown, integer()}

resolve_role_id(role_id)

resolve_role_id(role :: Crux.Structs.Role.t() | integer()) :: integer()

resolve_user_id(id)

resolve_user_id(
  user :: Crux.Structs.User.t() | Crux.Structs.Member.t() | integer()
) :: integer()