Crux v2019-11-19 15:49:43Z Crux.Structs.AuditLogEntry

Represents a Discord Audit Log Object

Types

event_name() (since 0.1.6)

event_name() ::
  :guild_update
  | :channel_create
  | :channel_update
  | :channel_delete
  | :channel_overwrite_create
  | :channel_overwrite_update
  | :channel_overwrite_delete
  | :member_kick
  | :member_prune
  | :member_ban_add
  | :member_ban_remove
  | :member_update
  | :member_role_update
  | :role_create
  | :role_update
  | :role_delete
  | :invite_create
  | :invite_update
  | :invite_delete
  | :webhook_create
  | :webhook_update
  | :webhook_delete
  | :emoji_create
  | :emoji_update
  | :emoji_delete
  | :message_delete

Union type of audit log event name atoms.

t() (since 0.1.6)

t() :: %Crux.Structs.AuditLogEntry{
  action_type: non_neg_integer(),
  changes: %{required(String.t()) => Crux.Structs.AuditLogChange.t()},
  id: Crux.Structs.Snowflake.t(),
  options: %{} | nil,
  reason: String.t() | nil,
  target_id: Crux.Structs.Snowflake.t(),
  user_id: Crux.Structs.Snowflake.t()
}

Functions

create(data) (since 0.1.6)

create(data :: map()) :: t()

Creates a Crux.Structs.AuditLogEntry.t/0 struct from raw data.

Automatically invoked by Crux.Structs.create/2.

event_name(action_type) (since 0.1.6)

event_name(action_type :: non_neg_integer()) :: atom()

Gets the event name from the action type id

events() (since 0.1.6)

events() :: %{required(event_name()) => non_neg_integer()}

Returns a map of all audit log event names with their id